% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{An examplar dataset.}
\format{A data frame with 1101 rows and 7 variables:
\describe{
  \item{thread}{name of thread}
  \item{grade}{grade level, 1-6}
  \item{id}{sequential id}
  \item{noteid}{note id}
  \item{coding}{codes, 6 categories}
  \item{type}{two types of threads, e and i}
  \item{unit}{study unit, not really relevant}
}}
\usage{
df
}
\description{
A dataset containing two groups (types) of threads. Each thread contains
varied number of coded entries (notes).
}
\keyword{datasets}

