% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LagSeq_function.R
\name{LagSeq}
\alias{LagSeq}
\title{A basic function that converts data to a transitional frequency matrix}
\usage{
LagSeq(vec, ncodes = 0, lag = 1, merge = FALSE)
}
\arguments{
\item{vec}{A vector of data representing a sequence}

\item{ncodes}{Integer. The number of codes (or types of events/behaviors) expected in the vector. Optional.}

\item{lag}{Integer. The lag to be applied to the conversion. Default: 1, denoting converting based on immediate adjacency.}

\item{merge}{Boolean. Whether to merge the same codes appearing without interruption. Default: FALSE.}
}
\value{
A list of a bunch of matrices representing different statistics of transitional frequency
}
\description{
This function reads a vector of data and then computes and returns a square output matrix (the cells of which correspond to the cells of the transitional frequency matrix) for each ofthe following: transitional frequencies with row and column totals, ex- pected frequencies, transitional probabilities, adjusted residuals and significance levels, Yule's Q values, transformed kappas (Wampold , 1989, 1992, 1995), z values for the kappas, and significance levels.
}
\examples{
vec = c(1, 2, 3, 4, 3, 2, 1, 2)
LagSeq(vec, ncodes = 4)
}

