% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LagSeq_function.R
\name{LagSeq_Groups}
\alias{LagSeq_Groups}
\title{Compare transitional patterns between two groups}
\usage{
LagSeq_Groups(df, group, seq, codes, measure = "freq", ncodes = 0,
  lag = 1, merge = FALSE)
}
\arguments{
\item{df}{A data frame containing required data. Data should be strict tabular format, with at least the following columns---group membership, sequence membership, codes.}

\item{group}{Index of the column representing group membership.}

\item{seq}{Index of the column representing sequence membership.}

\item{codes}{Index of the column representing codes.}

\item{measure}{A vector containing measures to compare between two groups. Possible values include `freq`, `adr`, and `yule`.}

\item{ncodes}{Integer. The number of codes (or types of events/behaviors) expected in the vector. Optional.}

\item{lag}{Integer. The lag to be applied to the conversion. Default: 1, denoting converting based on immediate adjacency.}

\item{merge}{Boolean. Whether to merge the same codes appearing without interruption. Default: FALSE.}
}
\value{
Nothing. But results of comparisons will be printed.
}
\description{
Given two groups, each of which contains multiple sequences of codes (or types of events/behaviors), compare these two groups whether there is any significant difference for each pair of codes for given trasational relationship measure(s).
}
\examples{
load("lagseq_example_data.Rdata")
Lag_Seq_Groups(df, group=6, seq=1, codes=5)
}

