% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LagSeq_function.R
\name{MergeSameCodes}
\alias{MergeSameCodes}
\title{A utility function for merging same codes appearing together}
\usage{
MergeSameCodes(vec)
}
\arguments{
\item{vec}{A vector of data representing a sequence}
}
\value{
The resulting vector after merging
}
\description{
This function reads a vector of data and then merges same codes that appear together (i.e., without being interrupted by other codes)
}
\examples{
vec = c(1, 2, 2, 3, 4)
MergeSameCodes(vec)
}

